Fertig, aber nicht getestet

+++ src/uebung5/Punkt.java
@@ -1,28 +1,28 @@
 package uebung5;
 
 public class Punkt {
-	private int x;
-	private int y;
+	private double x;
+	private double y;
 	
 	public Punkt() {
 		this.x = 0;
 		this.y = 0;
 	}
 	
-	public Punkt(int x, int y) {
+	public Punkt(double x, double y) {
 		this.x = x;
 		this.y = y;
 	}
 	
-	public int getX() {
+	public double getX() {
 		return x;
 	}
 	
-	public int getY() {
+	public double getY() {
 		return y;
 	}
 	
-	public void verschiebe(int x, int y) {
+	public void verschiebe(double x, double y) {
 		this.x += x;
 		this.y += y;
 	}
